Moved purity checking from dataflow to common.purity
