+++ framework/src/main/java/org/checkerframework/common/purity/qual/Deterministic.java
@@ -1,0 +1,3 @@
+package org.checkerframework.common.purity.qual;
+
+public @interface Deterministic {}
